va_claims_manager/requirements.txt
# VA Disability Claims Manager - Dependencies
# Install: pip install -r requirements.txt
# System dependency: Tesseract 5.x
#   winget install UB-Mannheim.TesseractOCR

# GUI
PyQt6==6.7.1

# PDF processing
pdfplumber==0.11.4
pypdf==4.3.1
reportlab==4.2.2
pdfrw==0.4

# Word document processing
python-docx==1.1.2

# OCR and image processing
pytesseract==0.3.13
Pillow==10.4.0
opencv-python-headless==4.10.0.84

# Utilities
python-dateutil==2.9.0.post0
rapidfuzz==3.9.7

# Testing
pytest==8.3.2
pytest-qt==4.4.0
